requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe requirements.in
#
aiohttp==3.9.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via
    #   requests
    #   yarl
influxdb==5.3.2
    # via -r requirements.in
msgpack==1.0.8
    # via influxdb
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pydantic==2.7.1
    # via -r requirements.in
pydantic-core==2.18.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via influxdb
pytz==2024.1
    # via influxdb
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   influxdb
six==1.16.0
    # via
    #   influxdb
    #   python-dateutil
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
yarl==1.9.4
    # via aiohttp
